fix: BOT_CATEGORIES 순환 의존성 제거 — signalBotCategories.js 분리
